# openapi.yaml
openapi: 3.0.0
info:
  title: 'My App'
  description: 'This is an awesome app!'
  version: 1.0.0
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: 'List users with pagination and sorting'
      operationId: get_api_users_list
      parameters:
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 0
            exclusiveMinimum: true
        -
          name: itemsPerPage
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 100
            minimum: 0
            exclusiveMinimum: true
        -
          name: orderBy
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        -
          name: sortField
          in: query
          required: false
          schema:
            type: string
            default: email
            enum:
              - email
              - created_at
              - id
      responses:
        '200':
          description: 'Returns paginated list of users'
          content:
            application/json:
              schema:
                properties:
                  users: { type: array, items: { properties: { id: { type: string, format: uuid, example: 550e8400-e29b-41d4-a716-446655440000 }, email: { type: string, format: email, example: user@example.com }, roles: { type: array, items: { type: string }, example: [ROLE_USER] }, is_2fa_enabled: { type: boolean, example: false }, created_at: { type: string, format: date-time, example: '2025-10-17T10:30:00+00:00' }, updated_at: { type: string, format: date-time, example: '2025-10-17T12:45:00+00:00', nullable: true } }, type: object } }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
                type: object
        '400':
          description: 'Invalid query parameters'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Validation failed' }
                type: object
      security:
        -
          Bearer: []
    post:
      tags:
        - Users
      summary: 'Create a new user'
      operationId: post_api_users_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDTO'
      responses:
        '201':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                properties:
                  user_id: { type: string, format: uuid, example: 550e8400-e29b-41d4-a716-446655440000 }
                  email: { type: string, format: email, example: user@example.com }
                  roles: { type: array, items: { type: string }, example: [ROLE_USER] }
                  is_two_factor_enabled: { type: boolean, example: false }
                type: object
        '400':
          description: 'Invalid request data'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Validation failed' }
                type: object
        '409':
          description: 'User already exists'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'User with email "user@example.com" already exists.' }
                type: object
      security:
        -
          Bearer: []
components:
  schemas:
    PaginationMeta:
      properties:
        total:
          description: 'Total number of items'
          type: integer
          example: 42
        page:
          description: 'Current page number'
          type: integer
          example: 1
        items_per_page:
          description: 'Number of items per page'
          type: integer
          example: 10
        total_pages:
          description: 'Total number of pages'
          type: integer
          example: 5
      type: object
...
    CreateUserDTO:
      required:
        - email
        - password
        - isTwoFactorEnabled
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        roles:
          type: array
          items:
            type: string
          default:
            - ROLE_USER
        isTwoFactorEnabled:
          type: boolean
        totpSecret:
          type: string
          default: null
          nullable: true
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []
...